+++ src/components/App/App.js
@@ -2,16 +2,15 @@ import LandingPage from "../LandingPage";
 import About from "../About";
 import React from "react";
 import { styled } from "@mui/material/styles";
-import { ThemeProvider } from "@mui/material/styles";
 import Box from "@mui/material/Box";
 import { Routes, Route } from "react-router-dom";
 import Navbar from "../Navbar";
 import CourseHome from "../CourseHome";
 import SectionPage from "../SectionPage";
 import course from "../../data/course.js";
-import theme from "../Theme";
 import { AddressProvider } from "../AddressContext";
 import { ContractProvider } from "../ContractContext";
+import {ThemeContext, themes} from '../ThemeContext';
 
 const AppContainer = styled(Box)(() => ({
   height: "100vh",
@@ -25,38 +24,42 @@ function App() {
   const [contractData, setContractData] = React.useState({});
   const contractValue = { contractData, setContractData };
 
-  return (
-    <ThemeProvider theme={theme}>
-      <AddressProvider value={value}>
-        <ContractProvider value={contractValue}>
-          <AppContainer id="app-container">
-            <Navbar />
+  const [customTheme, setCustomTheme] = React.useState("light")
+  const themeValue = { customTheme, setCustomTheme };
 
-            <Routes>
-              <Route exact path="/" element={<LandingPage />}></Route>
-              <Route path="/about" element={<About />}></Route>
-              <Route path="/home" element={<LandingPage />}></Route>
-              <Route path="/course/*" element={<CourseHome />}>
-                {course.chapters.map((chapter) =>
-                  chapter.sections.map((section) => (
-                    <Route
-                      path={`${chapter.url}/${section.url}`}
-                      element={
-                        <SectionPage
-                          contentUrl={section.contentUrl}
-                          hasCodeEditor={section.hasCodeEditor}
-                          defaultCode={section.defaultCode}
-                        />
-                      }
-                    />
-                  ))
-                )}
-              </Route>
-            </Routes>
-          </AppContainer>
-        </ContractProvider>
-      </AddressProvider>
-    </ThemeProvider>
+
+  return (
+    <ThemeContext.Provider value={themeValue}>
+        <AddressProvider value={value}>
+          <ContractProvider value={contractValue}>
+            <AppContainer id="app-container">
+              <Navbar />
+
+              <Routes>
+                <Route exact path="/" element={<LandingPage />}></Route>
+                <Route path="/about" element={<About />}></Route>
+                <Route path="/home" element={<LandingPage />}></Route>
+                <Route path="/course/*" element={<CourseHome />}>
+                  {course.chapters.map((chapter) =>
+                    chapter.sections.map((section) => (
+                      <Route
+                        path={`${chapter.url}/${section.url}`}
+                        element={
+                          <SectionPage
+                            contentUrl={section.contentUrl}
+                            hasCodeEditor={section.hasCodeEditor}
+                            defaultCode={section.defaultCode}
+                          />
+                        }
+                      />
+                    ))
+                  )}
+                </Route>
+              </Routes>
+            </AppContainer>
+          </ContractProvider>
+        </AddressProvider>
+    </ThemeContext.Provider>
   );
 }
