--- src/components/Theme/Theme.js
@@ -1,15 +1,0 @@
-import { createTheme } from "@mui/material/styles";
-import { orange, blue } from "@mui/material/colors";
-
-const Theme = createTheme({
-  palette: {
-    primary: {
-      main: blue[500],
-    },
-    secondary: {
-      main: orange[400],
-    },
-  },
-});
-
-export default Theme;
