--- src/components/Theme/index.js
@@ -1,2 +1,0 @@
-import Theme from "./Theme";
-export default Theme;
